pyproject.toml
[tool.poetry]
name = "thepeak-spotify"
version = "1.0.0"
description = "Automatic playlist generator for Spotify"
authors = ["William Abreu <william@t10.digital>"]

[tool.poetry.dependencies]
python = "^3.9"
spotipy = "^2.19.0"

[tool.poetry.dev-dependencies]
mypy = "^0.960"
pydocstyle = "^6.1.1"
pre-commit = "^2.19.0"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pyclean = "^2.2.0"
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# See: https://mypy.readthedocs.io/en/stable/config_file.html

[tool.mypy]
exclude = "(\\..*|env|venv|__pycache__)/$"
files = ["core", "serverless"]
python_version = "3.9"
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true
pretty = true
